backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
opencv-python-headless
numpy
Pillow
torch
torchvision
